Most of the Javascript is now Typescript
All that's remaining is the tricky part, renderer.js. Also to note theres a new bug where payloads won't launch in the full builds. I think this is because of path issues.
