+++ .gitignore
@@ -1,6 +1,6 @@
 # super-payload-launcher specific
 drivercheckcomplete
-payloads/
+#payloads/
 
 # Logs
 logs
@@ -91,3 +91,106 @@ typings/
 
 # Electron-Forge
 out/
+
+# TYPESCRIPT SPECIFIC
+
+node_modules/
+.node_modules/
+built/*
+tests/cases/rwc/*
+tests/cases/test262/*
+tests/cases/perf/*
+!tests/cases/webharness/compilerToString.js
+test-args.txt
+~*.docx
+\#*\#
+.\#*
+tests/baselines/local/*
+tests/baselines/local.old/*
+tests/services/baselines/local/*
+tests/baselines/prototyping/local/*
+tests/baselines/rwc/*
+tests/baselines/test262/*
+tests/baselines/reference/projectOutput/*
+tests/baselines/local/projectOutput/*
+tests/baselines/reference/testresults.tap
+tests/services/baselines/prototyping/local/*
+tests/services/browser/typescriptServices.js
+src/harness/*.js
+src/compiler/diagnosticInformationMap.generated.ts
+src/compiler/diagnosticMessages.generated.json
+src/parser/diagnosticInformationMap.generated.ts
+src/parser/diagnosticMessages.generated.json
+rwc-report.html
+*.swp
+build.json
+*.actual
+tests/webTestServer.js
+tests/webTestServer.js.map
+tests/webhost/*.d.ts
+tests/webhost/webtsc.js
+tests/cases/**/*.js
+!tests/cases/docker/*.js/
+tests/cases/**/*.js.map
+*.config
+scripts/eslint/built/
+scripts/debug.bat
+scripts/run.bat
+scripts/word2md.js
+scripts/buildProtocol.js
+scripts/ior.js
+scripts/authors.js
+scripts/configurePrerelease.js
+scripts/configureLanguageServiceBuild.js
+scripts/open-user-pr.js
+scripts/open-cherry-pick-pr.js
+scripts/processDiagnosticMessages.d.ts
+scripts/processDiagnosticMessages.js
+scripts/produceLKG.js
+scripts/importDefinitelyTypedTests/importDefinitelyTypedTests.js
+scripts/generateLocalizedDiagnosticMessages.js
+scripts/request-pr-review.js
+scripts/*.js.map
+scripts/typings/
+coverage/
+internal/
+**/.DS_Store
+.settings
+**/.vs
+**/.vscode/*
+!**/.vscode/tasks.json
+!**/.vscode/settings.template.json
+!**/.vscode/launch.template.json
+!**/.vscode/extensions.json
+!tests/cases/projects/projectOption/**/node_modules
+!tests/cases/projects/NodeModulesSearch/**/*
+!tests/baselines/reference/project/nodeModules*/**/*
+.idea
+yarn.lock
+yarn-error.log
+.parallelperf.*
+tests/cases/user/*/package-lock.json
+tests/cases/user/*/node_modules/
+tests/cases/user/*/**/*.js
+tests/cases/user/*/**/*.js.map
+tests/cases/user/*/**/*.d.ts
+!tests/cases/user/zone.js/
+!tests/cases/user/bignumber.js/
+!tests/cases/user/discord.js/
+tests/baselines/reference/dt
+.failed-tests
+TEST-results.xml
+package-lock.json
+tests/cases/user/npm/npm
+tests/cases/user/TypeScript-React-Starter/TypeScript-React-Starter
+tests/cases/user/TypeScript-Node-Starter/TypeScript-Node-Starter
+tests/cases/user/TypeScript-React-Native-Starter/TypeScript-React-Native-Starter
+tests/cases/user/TypeScript-Vue-Starter/TypeScript-Vue-Starter
+tests/cases/user/TypeScript-WeChat-Starter/TypeScript-WeChat-Starter
+tests/cases/user/create-react-app/create-react-app
+tests/cases/user/fp-ts/fp-ts
+tests/cases/user/webpack/webpack
+tests/cases/user/puppeteer/puppeteer
+tests/cases/user/axios-src/axios-src
+tests/cases/user/prettier/prettier
+.eslintcache
