+++ .vscode/launch.json
@@ -5,19 +5,30 @@
     "version": "0.2.0",
     "configurations": [
         {
-            "name": "Debug Main Process",
+            "name": "Debug application",
             "type": "node",
             "request": "launch",
             "env": {
                 "ELECTRON_ENABLE_LOGGING": "true"
             },
             "cwd": "${workspaceFolder}",
+            "preLaunchTask": "Build Typescript",
             "runtimeExecutable": "${workspaceFolder}/node_modules/.bin/electron",
             "windows": {
-              "runtimeExecutable": "${workspaceFolder}/node_modules/.bin/electron.cmd"
+                "runtimeExecutable": "${workspaceFolder}/node_modules/.bin/electron.cmd"
             },
-            "args" : ["."],
+            "args": [
+                "."
+            ],
             "outputCapture": "std"
+        },
+        {
+            "command": "npm run make",
+            "cwd": "${workspaceFolder}",
+            "preLaunchTask": "Build Typescript",
+            "name": "Build application release",
+            "request": "launch",
+            "type": "node-terminal"
         }
     ]
 }
