+++ package.json
@@ -3,7 +3,7 @@
   "productName": "Super Payload Launcher",
   "version": "1.2.0",
   "description": "An easy and user friendly Nintendo Switch payload injecting tool",
-  "main": "src/index.js",
+  "main": "dist/app.js",
   "scripts": {
     "start": "electron-forge start",
     "package": "electron-forge package",
@@ -23,7 +23,21 @@
   "config": {
     "forge": {
       "packagerConfig": {
-        "icon": "build/icon.ico"
+        "icon": "build/icon.ico",
+        "overwrite": true,
+        "ignore": [
+          "^(/.github$)",
+          "^(/build$)",
+          "^(/.vscode$)",
+          "^(/src$)",
+          "drivercheckcomplete",
+          ".gitignore",
+          "LICENSE",
+          "README.md",
+          "package-lock.json",
+          "tsconfig.json",
+          "spl.code-workspace"
+        ]
       },
       "makers": [
         {
@@ -65,6 +79,7 @@
     "@electron-forge/maker-squirrel": "^6.0.0-beta.54",
     "@electron-forge/maker-zip": "^6.0.0-beta.54",
     "electron": "11.3.0",
-    "electron-rebuild": "^2.3.5"
+    "electron-rebuild": "^2.3.5",
+    "typescript": "^4.2.3"
   }
 }
