+++ src/app.ts
@@ -1,0 +1,4 @@
+import { app, BrowserWindow } from 'electron';
+import Main from './main';
+
+Main.main(app, BrowserWindow);
